Move known coordinators list to StatusContainer
