+++ WalletWasabi.Client/Global.cs
@@ -102,8 +102,6 @@ public class Global
 		TransactionBroadcaster = new TransactionBroadcaster(broadcasters.ToArray(), mempoolService, WalletManager);
 
 		Scheme = new Scheme(this);
-
-		CoordinatorDirectory = CoordinatorDirectory.LoadBundled();
 	}
 
 	private readonly AsyncLock _initializationAsyncLock = new();
@@ -132,7 +130,6 @@ public class Global
 	public Uri? OnionServiceUri { get; private set; }
 	public EventBus EventBus { get; }
 	public Scheme Scheme { get; }
-	public CoordinatorDirectory CoordinatorDirectory { get; }
 
 	private string GetBitcoinP2pNetworkDirectory() => Path.Combine(DataDir, "BitcoinP2pNetwork");
