+++ WalletWasabi.Fluent/App.axaml.cs
@@ -151,7 +151,6 @@ public class App : Application
 			torStatusChecker,
 			new HealthMonitor(torStatusChecker),
 			new ReleaseHighlights(),
-			Services.CoordinatorDirectory,
 			Services.Scheme);
 	}
 }
