+++ WalletWasabi.Fluent/Services.cs
@@ -3,7 +3,6 @@ using System.Net.Http;
 using WalletWasabi.Blockchain.Blocks;
 using WalletWasabi.Blockchain.TransactionBroadcasting;
 using WalletWasabi.Daemon;
-using WalletWasabi.Discoverability;
 using WalletWasabi.Helpers;
 using WalletWasabi.Services;
 using WalletWasabi.Services.Terminate;
@@ -47,7 +46,6 @@ public static class Services
 	public static EventBus EventBus { get;  set; } = null!;
 
 	public static Daemon.Scheme Scheme { get; set; } = null!;
-	public static CoordinatorDirectory CoordinatorDirectory { get; private set; } = null!;
 	public static bool IsInitialized { get; private set; }
 
 	/// <summary>
@@ -81,7 +79,6 @@ public static class Services
 		EventBus = global.EventBus;
 		Status = global.Status;
 		Scheme = global.Scheme;
-		CoordinatorDirectory = global.CoordinatorDirectory;
 
 		IsInitialized = true;
 	}
